// src/store/customer/customer.enum.ts
enum CustomerActionTypeEnum {
  CUSTOMER_LOADING = "CUSTOMER_LOADING",
  CUSTOMER_LOADED = "CUSTOMER_LOADED",
  GET_CUSTOMER = "GET_CUSTOMER",
  EDIT_CUSTOMER = "EDIT_CUSTOMER",
  UPDATE_CUSTOMER = "UPDATE_CUSTOMER_SUCCESS",
  DELETE_CUSTOMER = "DELETE_CUSTOMER_SUCCESS",
  CUSTOMER_ID = "CUSTOMER_ID",
  GET_CUSTOMER_ORDERS_FUEL = "GET_CUSTOMER_ORDERS_FUEL",
  GET_CUSTOMER_ORDERS_TANK = "GET_CUSTOMER_ORDERS_TANK",
  GET_ALL_CUSTOMERS = "GET_ALL_CUSTOMERS",
  UPDATE_CUSTOMER_STATUS = "UPDATE_CUSTOMER_STATUS",
  GET_CUSTOMER_BY_ID = "GET_CUSTOMER_BY_ID",
  DELETE_CUSTOMERS = "DELETE_CUSTOMERS",
}

export default CustomerActionTypeEnum;